+++ examples/email_edit.py
@@ -7,10 +7,10 @@ But it have increased memory consumption: ~10%
 """
 
 import email
-from imap_tools import MailBox
+from imap_tools import MailBox, A
 
 with MailBox('imap.mail.com').login('test@mail.com', 'password') as mailbox:
-    for msg in mailbox.fetch():
+    for msg in mailbox.fetch(A(subject='[some]')):
         altered_msg = email.message_from_bytes(msg.obj.as_bytes())
         altered_msg.replace_header('Subject', '[my own subject extension] ' + msg.subject)
         mailbox.append(altered_msg.as_bytes())